Fix two bugs reported by dcoombs.
test-sh was assuming 'bup' was on the PATH.  (It wasn't *supposed* to be
assuming that, but the "alias bup=whatever" line wasn't working,
apparently.)

randomgen.c triggered a warning in some versions of gcc about the return
value of write() being ignored.  It really doesn't bother me if some of my
random bytes don't get written, but whatever; I'll assert instead, which
should shut it up.