client-server: only retrieve index files when actually needed.
A busy server could end up with a *large* number of index files, mostly
referring to objects from other clients.  Downloading all the indexes not only
wastes bandwidth, but causes a more insidious problem: small servers end up
having to mmap a huge number of large index files, which sucks lots of RAM.

In general, the RAM on a server is roughly proportional to the disk space on
that server.  So it's okay for larger clients to need more RAM in order
to complete a backup.  However, it's not okay for the existence of larger
clients to make smaller clients suffer.  Hopefully this change will settle
it a bit.