cmd-ls and cmd-fuse: toys for browsing your available backups.
'bup ls' lets you browse the set of backups on your current system.  It's a
bit useless, so it might go away or be rewritten eventually.

'bup fuse' is a simple read-only FUSE filesystem that lets you mount your
backup sets as a filesystem (on Linux only).  You can then export this
filesystem over samba or NFS or whatever, and people will be able to restore
their own files from backups.

Warning: we still don't support file metadata in 'bup save', so all the file
permissions will be wrong (and users will probably be able to see things
they shouldn't!).  Also, anything that has been split into chunks will show
you the chunks instead of the full file, which is a bit silly.  There are
also tons of places where performance could be improved.

But it's a pretty neat toy nevertheless.  To try it out:

   mkdir /tmp/backups
   sudo bup fuse /tmp/backups