Another suspicious fix for CatPipe parallelism.
This really shouldn't be necessary: it's clear to me that the 'it' object
should be going out of scope right away, and thus getting cleaned up by the
garbage collector.

But on one of my Linux PCs (with python 2.4.4) it fails the unit tests
unless I add this patch.  Oh well, let's do it then.