main: list common commands before other ones.
When you just type 'bup' or 'bup help', we print a list of available
commands.  Now we improve this list by:

1) Listing the common commands (with one-line descriptions) before listing
the automatically-generated list.

2) Printing the automatically-generated list in columns, so it takes up less
vertical space.

This whole concept was stolen from how git does it.  I think it should be a
bit more user friendly for beginners this way.