bup random: fix progress output and don't print to a tty.
We were printing output using a series of dots, which interacted badly with
bup newliner (and for good reason).  Change it to actually display the
number of megabytes done so far.

Also, don't print random binary data to a tty unless -f is given.  It's
just more polite that way.