'make stupid' stopped working when I moved subcommands into their own dir.
Remote server mode tries to add the directory of argv[0] (the
currently-running program) to the PATH on the remote server, just in case
bup isn't installed in the PATH there, so that it can then run 'bup server'.

However, now that bup-save is in a different place than bup, argv[0] is the
wrong place to look.  Instead, have the bup executable export an environment
variable containing its location, and client.py can use that instead of
argv[0].  Slightly gross, but it works.