code clarity: one-letter var carried for too long
In split-cmd.py, the "w" variable is first seen on line 55 and is kept
around until line 96. Variables that are sparsely used in a medium
distance in the code should have a name that carries more sense when
read on its own.

Change "w" into "pack_writer" to better identify the purpose of the
variable.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>