Update tornado to revision ad104ffb41
The file lib/tornado/escape.py was forcing users to install a json
library even though "bup web" doesn't use any json functionality.

An issue was opened upstream:

http://github.com/facebook/tornado/issues/closed#issue/114

and the day after it was opened, a fix was committed for it.

Update to the latest revision of tornado so that we can remove a
dependency on json/simplejson.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>