cmd/ftp: if completion fails due to FileNotFound, just eat it.
Just as bash would do, if you're trying to complete a filename that doesn't
exist, just don't offer any completions.  In this case, it only happens if
you try to complete through a broken symlink.

Now that we've fixed this case, enable the printing of exception tracebacks
in case of *other* kinds of completion errors, since we don't expect there
to be any.

[Committed by apenwarr based on an unofficial patch from Gabriel]