cmd/web: hide .dotfiles by default
Make all files begining with a dot be hidden by default. The hidden
files can be shown by giving the argument "hidden" with a vlue of 1 in
the URL.

Also, in _compute_dir_contents, remove the line "contents = []" since it
is never used.

Finally add a "Show/Hide hidden files" link on the pages where content
is hidden.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>