cmd/midx: fix SHA_PER_PAGE calculation.
For some reason we were dividing by 200 instead of by 20, which was way off.
Switch to 20 instead.  Suspiciously, this makes memory usage slightly worse
in my current (smallish) set of test data, so we might need to revert it
later...?  But if we're going to have an adjustment, we should at least make
it clear what for, rather than hiding it in something that looks
suspiciously like a typo.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>