Rename Documentation/*.1.md to Documentation/*.md
All our man pages end up in section 1 of man anyway, and it looks like that
will probably never change.  So let's make our filenames simpler and easier
to understand.

Even if we do end up adding a page in (say) section 5 someday, it's no big
deal; we can just add an exception to the Makefile for it or something.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>