options.py: handle optspecs that include inline square brackets.
We recently made it so if the last thing on an options line was [defval],
then the value in brackets became the default for that option.  However, we
inadvertently matched *any* bracketed value on that line, not just the one
at the end of the line, which basically prevents us from using square
brackets anywhere on the line.  That's no fun.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>