options.py: remove extra newlines in usage string.
If the first line after the "--" was a comment (started with whitespace),
then we'd end up printing a double newline instead of a single one after the
synopsis string.

It would also look weird if we had a multi-line comment; the lines would be
separated by blank lines.

'bup damage -?' encountered this problem.

Signed-off-by: Avery Pennarun <apenwarr@gmail.com>