--remote parameter requires a colon
This patch checks for the presence of a colon if the --remote option
is used in bup save, bup split, bup join, and bup init.  Even though
specifying *only* a pathname without a hostname: is perfectly valid,
it's confusing to allow users to do so, because if they specify
"-r hostname" it will be treated as a path and thus give them a
confusing error message. Requiring a colon will avoid this.

It adds a few test cases to demonstrate that the code
works properly.

It also wraps the remote connection in a try except to prevent
a traceback if there is an error (so far I have only seen this
happen with an invalid bup dir parameter)

And I added the netbeans project folder to gitignore

Signed-off-by: David Roda <davidcroda@gmail.com>