Enable static resources. Move css to external file.
Add to the settings variable in web-cmd.py to set
/web/static to be servced as static resources.  This is for
css, javascript, and images.

Move the current styles from the head to static/css/styles.css.
Remove a few unnecessary styles and change the tab stop
to 4 spaces to match the rest of the code.

Add to Makefile to copy new directory structure.

Signed-off-by: David Roda <davidcroda@gmail.com>